# [ray] refactor: Accelerate Tensor serialization by converting to np.ndarray