--- a/verl/protocol.py
+++ b/verl/protocol.py
@@ -35,7 +35,7 @@ from packaging.version import parse as parse_version
 from tensordict import TensorDict
 from torch.utils.data import DataLoader
 
-from verl.utils.device import get_device_id, get_torch_device
+from verl.utils.device import get_device_id
 from verl.utils.py_functional import union_two_dict
 from verl.utils.torch_functional import allgather_dict_tensors