### What does this PR do?

For a data size of 6400x20480, the average serialization duration was reduced from 3.32s to 1.32s following this optimization, resulting in a ~151% improvement.

```
# tensor
average  serialize：2.58s  deserialize：0.74s  total：3.32s

TaskRunner pid=1904793) baymax debug serialize time=2.5947s
(TaskRunner pid=1904793) baymax debug serialize time=2.593357s
(TaskRunner pid=1904793) baymax debug serialize time=2.580081s
(TaskRunner pid=1904793) baymax debug serialize time=2.582321s
(WorkerDict pid=1905183) baymax debug deserialize time=0.475745s
(WorkerDict pid=1905184) baymax debug deserialize time=0.538223s
(WorkerDict pid=1905181) baymax debug deserialize time=0.609146s
(WorkerDict pid=1905182) baymax debug deserialize time=0.61064s
(WorkerDict pid=1905189) baymax debug deserialize time=0.597746s
(WorkerDict pid=1905185) baymax debug deserialize time=0.530353s
(WorkerDict pid=1905180) baymax debug deserialize time=0.811555s
(WorkerDict pid=1905194) baymax debug deserialize time=0.513646s
(WorkerDict pid=1905193) baymax debug deserialize time=0.962868s
(WorkerDict pid=1905179) baymax debug deserialize time=0.929226s
(WorkerDict pid=1905186) baymax debug deserialize time=0.701976s
(WorkerDict pid=1905191) baymax debug deserialize time=0.867236s
(WorkerDict pid=1905192) baymax debug deserialize time=0.858472s
(WorkerDict pid=1905187) baymax debug deserialize time=1.045251s
(WorkerDict pid=1905188) baymax debug deserialize time=0.960867s
(WorkerDict pid=1905190) baymax debug deserialize time=1.010673s

# numpy
average  serialize：0.000617s  deserialize：1.32s  total：1.32s

[36m(TaskRunner pid=1729638)[0m baymax debug serialize time=0.00016s
[36m(TaskRunner pid=1729638)[0m baymax debug serialize time=0.000117s
[36m(TaskRunner pid=1729638)[0m baymax debug serialize time=0.000158s
[36m(TaskRunner pid=1729638)[0m baymax debug serialize time=0.000182s
[36m(WorkerDict pid=1730035)[0m baymax debug deserialize time=0.867232s
[36m(WorkerDict pid=1730036)[0m baymax debug deserialize time=0.97372s
[36m(WorkerDict pid=1730028)[0m baymax debug deserialize time=1.08627s
[36m(WorkerDict pid=1730034)[0m baymax debug deserialize time=1.187599s
[36m(WorkerDict pid=1730037)[0m baymax debug deserialize time=1.165926s
[36m(WorkerDict pid=1730025)[0m baymax debug deserialize time=1.281101s
[36m(WorkerDict pid=1730029)[0m baymax debug deserialize time=1.359834s
[36m(WorkerDict pid=1730027)[0m baymax debug deserialize time=1.281978s
[36m(WorkerDict pid=1730030)[0m baymax debug deserialize time=1.329298s
[36m(WorkerDict pid=1730026)[0m baymax debug deserialize time=1.475415s
[36m(WorkerDict pid=1730031)[0m baymax debug deserialize time=1.422345s
[36m(WorkerDict pid=1730033)[0m baymax debug deserialize time=1.378894s
[36m(WorkerDict pid=1730039)[0m baymax debug deserialize time=1.368721s
[36m(WorkerDict pid=1730040)[0m baymax debug deserialize time=1.601587s
[36m(WorkerDict pid=1730042)[0m baymax debug deserialize time=1.768378s
[36m(WorkerDict pid=1730038)[0m baymax debug deserialize time=1.765994s
```

### Checklist Before Starting

- [x] Search for similar PRs. Paste at least one query link here: ...
- [x] Format the PR title as `[{modules}] {type}: {description}` (This will be checked by the CI)
  - `{modules}` include `fsdp`, `megatron`, `sglang`, `vllm`, `rollout`, `trainer`, `ci`, `training_utils`, `recipe`, `hardware`, `deployment`, `ray`, `worker`, `single_controller`, `misc`, `perf`, `model`, `algo`, `env`, `tool`, `ckpt`, `doc`, `data`
  - If this PR involves multiple modules, separate them with `,` like `[megatron, fsdp, doc]`
  - `{type}` is in `feat`, `fix`, `refactor`, `chore`, `test`
  - If this PR breaks any API (CLI arguments, config, function signature, etc.), add `[BREAKING]` to the beginning of the title.
  - Example: `[BREAKING][fsdp, megatron] feat: dynamic batching`

### Test

> For changes that can not be tested by CI (e.g., algorithm implementation, new model support), validate by experiment(s) and show results like training curve plots, evaluation results, etc.

### API and Usage Example

> Demonstrate how the API changes if any, and provide usage example(s) if possible.

```python
# Add code snippet or script demonstrating how to use this
```

### Design & Code Changes

> Demonstrate the high-level design if this PR is complex, and list the specific changes.

### Checklist Before Submitting

> [!IMPORTANT]
> Please check all the following items before requesting a review, otherwise the reviewer might deprioritize this PR for review.

- [x] Read the [Contribute Guide](https://github.com/volcengine/verl/blob/main/CONTRIBUTING.md).
- [x] Apply [pre-commit checks](https://github.com/volcengine/verl/blob/main/CONTRIBUTING.md#code-linting-and-formatting): `pre-commit install && pre-commit run --all-files --show-diff-on-failure --color=always`
- [ ] Add / Update [the documentation](https://github.com/volcengine/verl/tree/main/docs).
- [ ] Add unit or end-to-end test(s) to [the CI workflow](https://github.com/volcengine/verl/tree/main/.github/workflows) to cover all the code. If not feasible, explain why: ...
- [ ] Once your PR is ready for CI, send a message in [the `ci-request` channel](https://verl-project.slack.com/archives/C091TCESWB1) in [the `verl` Slack workspace](https://join.slack.com/t/verl-project/shared_invite/zt-3855yhg8g-CTkqXu~hKojPCmo7k_yXTQ). (If not accessible, please try [the Feishu group (飞书群)](https://applink.larkoffice.com/client/chat/chatter/add_by_link?link_token=772jd4f1-cd91-441e-a820-498c6614126a).)