@vermouth1992 2025-09-10
Could you create a test case for nested tensor? Thanks!  https://docs.pytorch.org/docs/stable/nested.html